Scraping: Move plugin interfaces to Extensibility subnamespace
