+++ src/Snowflake.Framework.Primitives/Scraping/Extensibility/SeedTree.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace Snowflake.Scraping
+namespace Snowflake.Scraping.Extensibility
 {
     public struct SeedTree
     {
